requirements.txt
Flask~=2.3.0
Flask-RESTful~=0.3.9
Flask-SocketIO~=5.3.0
python-engineio~=4.3.0
python-socketio~=5.8.0
PyQt5~=5.15.0
PyQt5-tools~=5.15.0
SpeechRecognition~=3.10.0
pydub~=0.25.0
opencv-python~=4.8.0
Pillow~=10.0.0
requests~=2.31.0
beautifulsoup4~=4.12.0
scikit-learn~=1.3.0
numpy~=1.25.0
pandas~=2.0.0
matplotlib~=3.7.0
Flasgger~=0.9.7
gTTS~=2.3.0
pycryptodome~=3.19.0
langdetect~=1.0.9
pygettext~=1.0
polib~=1.1.0
python-dotenv~=1.0.0
SQLAlchemy~=1.4.0
psycopg2-binary~=2.9.0 # Opcional, si se usa PostgreSQL
Click~=8.1.0 # Para la CLI
Pydantic~=1.10.0 # Para validación de esquemas
python-multipart~=0.0.6 # Para Flasgger con Pydantic
cryptography~=41.0.0 # Para Fernet encryption
passlib~=1.7.0 # Para hashing de contraseñas
pyjwt~=2.8.0 # Para JWT
